+++ packages/client/hmi-client/src/components/workflow/tera-simulate-chart.vue
@@ -26,9 +26,9 @@ import _ from 'lodash';
 import { ref, computed, watch, onMounted } from 'vue';
 import MultiSelect from 'primevue/multiselect';
 import Chart from 'primevue/chart';
-import { ChartConfig, RunResults } from '@/types/SimulateConfig';
+import { ChartConfig, DataseriesConfig, RunResults, RunType } from '@/types/SimulateConfig';
 import { CsvAsset } from '@/types/Types';
-import { getGraphDataFromDatasetCSV, RunType, DatasetType } from './util';
+import { getGraphDataFromDatasetCSV } from './util';
 
 const emit = defineEmits(['configuration-change']);
 
@@ -206,12 +206,12 @@ const renderGraph = () => {
 		return;
 	}
 
-	const datasets: DatasetType[] = [];
+	const datasets: DataseriesConfig[] = [];
 	selectedVariable.value.forEach((variable) => {
 		runIdList
 			.map((runId) => renderedRuns.value[runId])
 			.forEach((run, runIdx) => {
-				const dataset: DatasetType = {
+				const dataset: DataseriesConfig = {
 					data: run.map(
 						// - runResults[selectedRun.value][timeIdx][code]
 						(datum: { [key: string]: number }) =>
@@ -230,7 +230,7 @@ const renderGraph = () => {
 			});
 
 		if (props.initialData) {
-			const dataset: DatasetType | null = getGraphDataFromDatasetCSV(
+			const dataset: DataseriesConfig | null = getGraphDataFromDatasetCSV(
 				props.initialData,
 				variable,
 				props.mapping,
