+++ packages/client/hmi-client/src/types/SimulateConfig.ts
@@ -1,3 +1,8 @@
+export const enum RunType {
+	Julia = 'julia',
+	Ciemss = 'ciemss'
+}
+
 export enum TspanUnits {
 	'Units',
 	'Date'
@@ -6,6 +11,15 @@ export enum TspanUnits {
 export type ChartConfig = {
 	selectedVariable: string[];
 	selectedRun: string;
+};
+
+export type DataseriesConfig = {
+	data: { x: number; y: number }[];
+	label: string;
+	fill: boolean;
+	borderColor?: string;
+	borderWidth?: number;
+	borderDash?: number[];
 };
 
 export type RunResults = {
